+++ app/Models/Media.php
@@ -7,7 +7,7 @@ namespace Modules\CMS\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Support\Carbon;
 use Modules\Core\Helpers\HasVersions;
-use Modules\Core\Helpers\SoftDeletes;
+use Modules\Core\SoftDeletes\SoftDeletes;
 use Override;
 use Spatie\MediaLibrary\MediaCollections\Models\Media as BaseMedia;